Refactor APIs into separate classes
